fix hidden movies after actors deletion

+++ models/Movie.js
@@ -196,14 +196,16 @@ export default class Movie {
 			],
 		);
 
-		await db.run(
-			`DELETE FROM Actors WHERE id in (SELECT actorId FROM MoviesActors where movieId=?)`,
-			[id],
-		);
+		if (actorsList && actorsList.length) {
+			await db.run(
+				`DELETE FROM Actors WHERE id in (SELECT actorId FROM MoviesActors where movieId=?)`,
+				[id],
+			);
 
-		await db.run(`DELETE FROM MoviesActors WHERE movieId=?`, [id]);
+			await db.run(`DELETE FROM MoviesActors WHERE movieId=?`, [id]);
 
-		await Movie.linkActorsToMovies(id, actorsList);
+			await Movie.linkActorsToMovies(id, actorsList);
+		}
 	}
 
 	static async deleteMovie(movieId) {
